Engine.SDL.Exception

Extensible Exceptions

newtype SDLException

This is thrown in the event of an error in the Engine.SDL combinators

Constructors

SDLException String

Utilities

err :: CInt -> IO ()

Treat negative return codes as prompting an error check.