Engine.SDL.Video

Attribute StateVars

contextMajorVersion :: Varied m => m Int

get/set SDL_GL_CONTEXT_MAJOR_VERSION, OpenGL context major version; see Remarks for details

contextMinorVersion :: Varied m => m Int

get/set SDL_GL_CONTEXT_MINOR_VERSION, OpenGL context major version; see Remarks for details

contextFlags :: Varied m => m Int

get/set SDL_GL_CONTEXT_FLAGS, some bitwise (.|.) of 0 or more of:

glContextFlagDebug

This flag maps to GLX_CONTEXT_DEBUG_BIT_ARB in the GLX_ARB_create_context extension for X11 and WGL_CONTEXT_DEBUG_BIT_ARB in the WGL_ARB_create_context extension for Windows. This flag is currently ignored on other targets that don't support equivalent functionality. This flag is intended to put the GL into a "debug" mode which might offer better developer insights, possibly at a loss of performance (although a given GL implementation may or may not do anything differently in the presence of this flag).

glContextFlagForwardCompatible

This flag maps to GLX_CONTEXT_FORWARD_COMPATIBLE_BIT_ARB in the GLX_ARB_create_context extension for X11 and WGL_CONTEXT_FORWARD_COMPATIBLE_BIT_ARB in the WGL_ARB_create_context extension for Windows. This flag is currently ignored on other targets that don't support equivalent functionality. This flag is intended to put the GL into a "forward compatible" mode, which means that no deprecated functionality will be supported, possibly at a gain in performance, and only applies to GL 3.0 and later contexts.

glContextFlagResetIsolation

This flag maps to GLX_CONTEXT_RESET_ISOLATION_BIT_ARB in the GLX_ARB_robustness_isolation extension for X11 and WGL_CONTEXT_RESET_ISOLATION_BIT_ARB in the WGL_ARB_create_context_robustness extension for Windows. This flag is currently ignored on other targets that don't support equivalent functionality. This flag is intended to require the GL to make promises about what to do in the face of driver or hardware failure.

glContextFlagRobustAccess

This flag maps to GLX_CONTEXT_ROBUST_ACCESS_BIT_ARB in the GLX_ARB_create_context_robustness extension for X11 and WGL_CONTEXT_ROBUST_ACCESS_BIT_ARB in the WGL_ARB_create_context_robustness extension for Windows. This flag is currently ignored on other targets that don't support equivalent functionality. This flag is intended to require a GL context that supports the GL_ARB_robustness extension--a mode that offers a few APIs that are safer than the usual defaults (think snprintf() vs sprintf()).

contextProfileMask :: Varied m => m Int

get/set SDL_GL_CONTEXT_PROFILE_MASK, which must be _one_ of

glProfileCore

glProfileCompatibility

glProfileES

Despite the name implying you could (.|.) these together, these are mutually exclusive!

redSize :: Varied m => m Int

get/set SDL_GL_RED_SIZE, the minimum number of bits for the red channel of the color buffer; defaults to 3

greenSize :: Varied m => m Int

get/set SDL_GL_GREEN_SIZE, the minimum number of bits for the green channel of the color buffer; defaults to 3

blueSize :: Varied m => m Int

get/set SDL_GL_BLUE_SIZE, the minimum number of bits for the blue channel of the color buffer; defaults to 2

alphaSize :: Varied m => m Int

get/set SDL_GL_ALPHA_SIZE, the minimum number of bits for the alpha channel of the color buffer; defaults to 0

bufferSize :: Varied m => m Int

get/set SDL_GL_BUFFER_SIZE, the minimum number of bits for frame buffer size; defaults to 0

depthSize :: Varied m => m Int

get/set SDL_GL_DEPTH_SIZE, the minimum number of bits in the depth buffer; defaults to 16

stencilSize :: Varied m => m Int

get/set SDL_GL_STENCIL_SIZE, the minimum number of bits in the stencil buffer; defaults to 0

accumRedSize :: Varied m => m Int

get/set SDL_GL_ACCUM_RED_SIZE, the minimum number of bits for the red channel of the accumulation buffer; defaults to 0

accumGreenSize :: Varied m => m Int

get/set SDL_GL_ACCUM_GREEN_SIZE, the minimum number of bits for the green channel of the accumulation buffer; defaults to 0

accumBlueSize :: Varied m => m Int

get/set SDL_GL_ACCUM_BLUE_SIZE, the minimum number of bits for the blue channel of the accumulation buffer; defaults to 0

accumAlphaSize :: Varied m => m Int

get/set SDL_GL_ACCUM_ALPHA_SIZE, the minimum number of bits for the alpha channel of the accumulation buffer; defaults to 0

multiSampleBuffers :: Varied m => m Int

get/set SDL_GL_MULTISAMPLEBUFFERS, the number of buffers used for multisample anti-aliasing; defaults to 0; see Remarks for details

multiSampleSamples :: Varied m => m Int

get/set SDL_GL_MULTISAMPLESAMPLES, the number of samples used around the current pixel used for multisample anti-aliasing; defaults to 0; see Remarks for details

stereo :: Varied m => m Bool

get/set SDL_GL_STEREO, whether the output is stereo 3D; defaults to off

acceleratedVisual :: Varied m => m Bool

get/set SDL_GL_ACCELERATED_VISUAL, set to True to require hardware acceleration, set to False to force software rendering; defaults to allow either

doubleBuffer :: Varied m => m Bool

get/set SDL_GL_DOUBLEBUFFER

shareWithCurrentContext :: Varied m => m Bool

get/set SDL_GL_SHARE_WITH_CURRENT_CONTEXT

framebufferSRGBCapable :: Varied m => m Bool

get/set SDL_GL_FRAMEBUFFER_SRGB_CAPABLE

swapInterval :: Varied m => m Int

Abstracts over SDL_GL_GetSwapInterval / SDL_GL_SetSwapInterval

0 for immediate updates, 1 for updates synchronized with the vertical retrace. -1 for late swap tearing (if supported) late swap tearing support can be checked under the GLX_EXT_swap_control_tear extension

windowDisplayMode :: Varied m => Window -> m DisplayMode

desktopDisplayMode :: MonadIO m => Int -> m DisplayMode